Add DXT build, improve tools for Claude 4
