+++ .github/workflows/build-dxt.yml
@@ -1,0 +1,101 @@
+name: Build and Test DXT Package
+
+on:
+  push:
+    branches: [ main, develop ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  build-and-test:
+    runs-on: ubuntu-latest
+    
+    steps:
+      - uses: actions/checkout@v4
+      
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '18'
+          cache: 'npm'
+      
+      - name: Install dependencies
+        run: npm ci
+      
+      - name: Run linting
+        run: npm run lint
+      
+      - name: Run unit tests
+        run: npm run test
+      
+      - name: Run e2e tests
+        run: npm run test:e2e
+      
+      - name: Build TypeScript
+        run: npm run build
+      
+      - name: Install DXT CLI
+        run: npm install -g @anthropic-ai/dxt
+      
+      - name: Build DXT package
+        run: |
+          chmod +x build-dxt.sh
+          ./build-dxt.sh
+      
+      - name: Upload DXT package artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: dxt-package
+          path: "*.dxt"
+          retention-days: 30
+      
+      - name: Verify DXT package
+        run: |
+          if [ -f *.dxt ]; then
+            echo "DXT package created successfully:"
+            ls -la *.dxt
+            file *.dxt
+          else
+            echo "No DXT package found"
+            exit 1
+          fi
+
+  test-dxt-install:
+    needs: build-and-test
+    runs-on: ubuntu-latest
+    
+    steps:
+      - name: Download DXT package
+        uses: actions/download-artifact@v4
+        with:
+          name: dxt-package
+      
+      - name: Test DXT package structure
+        run: |
+          # Extract and verify DXT package contents
+          for dxt_file in *.dxt; do
+            if [ -f "$dxt_file" ]; then
+              echo "Examining DXT package: $dxt_file"
+              unzip -l "$dxt_file" || echo "Could not list contents of $dxt_file"
+              
+              # Check for required files
+              unzip -q "$dxt_file" -d extracted/
+              
+              if [ -f "extracted/manifest.json" ]; then
+                echo "✓ manifest.json found"
+                cat extracted/manifest.json | head -20
+              else
+                echo "✗ manifest.json missing"
+                exit 1
+              fi
+              
+              if [ -f "extracted/dist/index.js" ]; then
+                echo "✓ dist/index.js found"
+              else
+                echo "✗ dist/index.js missing"
+                exit 1
+              fi
+              
+              echo "DXT package validation successful"
+            fi
+          done
