+++ .gitignore
@@ -120,3 +120,4 @@ dist
 # built artifacts
 build/
 dist/
+computer-use-mcp.dxt
