+++ package.json
@@ -19,9 +19,12 @@
     "start": "npm run build && node dist/index.js",
     "test": "vitest run",
     "test:watch": "vitest --watch",
+    "test:e2e": "vitest run --config vitest.e2e.config.ts",
     "lint": "eslint",
     "clean": "rm -rf dist",
     "build": "tsc --project tsconfig.build.json",
+    "build-dxt": "./build-dxt.sh",
+    "pack": "dxt pack",
     "prepublishOnly": "npm run clean && npm run build"
   },
   "dependencies": {
@@ -32,6 +35,7 @@
     "zod": "^3.24.1"
   },
   "devDependencies": {
+    "@anthropic-ai/dxt": "^0.1.0",
     "@tsconfig/node-lts": "^22.0.2",
     "@types/imagemin": "^9.0.0",
     "@types/node": "^22.10.2",
